pc_receiver/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
pynput==1.7.7
mss==9.0.1
Pillow==10.4.0
